refactor(cff): extract dict module for DICT parsing/serialization
Move TopDictOffsets, FdData, and all DICT/charset/FDSelect parsing and
serialization functions from the monolithic cff_subsetter.rs into a new
cff/dict.rs module. This is the third of three modularization tasks.
